refactor common printing
